Add support for ood-portal-generator
Apache's ood-portal.conf is currently managed exclusively by ansible
template, without the possibility of using ood-portal-generator

This commit add the update ood portal handler to generate
ood-portal.conf based on ood-portal.yml

* Add httpd htcacheclean handler
* Add ignore error com portal generation
