+++ molecule/default/tests/test_default.py
@@ -16,19 +16,6 @@ def test_httpd_running(host):
     assert len(httpd_process) > 0
 
 
-def test_ood_portal_conf(host):
-    hostname = host.check_output('hostname -s')
-    ood_portal_conf = '/opt/rh/httpd24/root/etc/httpd/conf.d/ood-portal.conf'
-
-    if hostname == 'custom':
-        header = '# Generated using ood-portal-generator version'
-    elif hostname == 'instance':
-        header = '# Ansible managed:'
-
-    # By default use ood-portal template
-    assert host.file(ood_portal_conf).contains(header)
-
-
 def test_httpd_running_and_enabled(host):
     httpd = host.service("httpd24-httpd")
     assert httpd.is_running
